Rotate coordinate system by 90° counter-clockwise, fixes #249
We rotate the coordinate system by 90 degrees counter-clockwise
so that when one looks onto the camera from the telescope dish,
x points right and y points up.

This is the same coordinate system used by MARS and makes
it easier to plot things into the camera, as x and y
fit intuition.

This is a breaking change that requires changes in all
downstream software handling coordinate transformations (e.g. pyfact)